requirements.txt
requests
python-dotenv
tqdm
yt-dlp
